requirements.txt
fastapi
uvicorn[standard]
pymongo
motor
pydantic
pydantic-settings
python-dotenv
httpx
ollama